Translator: Make vec/matrix size getters unsigned
TType stored type's primary and secondary sizes as `unsigned char`, but
the getters returned `int`.  This caused unnecessary casts when the size
was passed from one TType to another, as well as comparisons with other
unsigned numbers.

This change specifies the type of these members as `uint8_t` and makes
the getters return the same type.  The call sites are accordingly
adjusted to remove unnecessary casts, use the correct type in local
variables, and add casts when passed to ostream::operator<<.

Bug: angleproject:6755

+++ src/compiler/translator/TranslatorMetalDirect/Pipeline.cpp
@@ -303,8 +303,8 @@ static bool CompareBy(Compare op, const T &x, const T &y)
     }
 }
 
-template <TBasicType BT, Compare Cmp, int MatchDim, int NewDim>
-static int SaturateVectorOf(const TField &field)
+template <TBasicType BT, Compare Cmp, uint8_t MatchDim, uint8_t NewDim>
+static uint8_t SaturateVectorOf(const TField &field)
 {
     static_assert(NewDim >= MatchDim, "");
 
@@ -362,15 +362,15 @@ ModifyStructConfig Pipeline::externalStructModifyConfig() const
                 {
                     return 1;
                 }
-                if (int s = SaturateVectorOf<TBasicType::EbtInt, LT, 4, 4>(field))
+                if (uint8_t s = SaturateVectorOf<TBasicType::EbtInt, LT, 4, 4>(field))
                 {
                     return s;
                 }
-                if (int s = SaturateVectorOf<TBasicType::EbtUInt, LT, 4, 4>(field))
+                if (uint8_t s = SaturateVectorOf<TBasicType::EbtUInt, LT, 4, 4>(field))
                 {
                     return s;
                 }
-                if (int s = SaturateVectorOf<TBasicType::EbtFloat, LT, 4, 4>(field))
+                if (uint8_t s = SaturateVectorOf<TBasicType::EbtFloat, LT, 4, 4>(field))
                 {
                     return s;
                 }
